refactor: Move file and buildNotificationData func
To homogenize the place of notifications
